GET /api/workspace/issue/description: This API fetches the description of a specific issue. It uses the getIssue controller function which retrieves the issue details from the database.
PUT /api/workspace/issue/updateDescription: This API updates the description of a specific issue. It uses the updateIssueDescription controller function which updates the issue details in the database.

GET /api/workspace/issue/details: This API fetches the details of a specific issue, including the assignee details such as name of assignee and name of project which associated with specific issue. It uses the assigneDetails controller function which retrieves the issue and assignee details from the database.

POST /api/workspace/issue/comment: This API adds a new comment to a specific issue. It uses the addComment controller function which creates a new comment in the database and associates it with the specified issue.

GET /api/workspace/issue/comments: This API fetches the comments of a specific issue. It uses the getComments controller function which retrieves the comments associated with the specified issue from the database.